change project...

+++ db/migrate/20121113145355_create_likes.rb
@@ -1,8 +1,8 @@
 class CreateLikes < ActiveRecord::Migration
   def change
     create_table :likes do |t|
-      t.float :user_id
-      t.float :picture_id
+      t.integer :user_id
+      t.integer :picture_id
 
       t.timestamps
     end
